+++ R/wrapper_functions.R
@@ -142,7 +142,7 @@ newCommunityParams <- function(max_w = 1e6,
                        dim = c(1, length(params@w)))
     params <- setInitialValues(params, initial_n = initial_n)
 
-    params@rates_funcs$RDD <- "srrConstant"
+    params@rates_funcs$RDD <- "constantRDD"
     if (missing(recruitment)) {
         recruitment <- get_required_recruitment(params)
     }
@@ -1117,7 +1117,7 @@ retuneReproductionEfficiency <- function(params,
         }
     }
     params@species_params$erepro <- eff
-    return(setReproduction(params, srr = "srrNone"))
+    return(setReproduction(params, RDD = "noRDD"))
 }
 
 #' Determine recruitment rate needed for initial egg abundance
@@ -1158,7 +1158,7 @@ setRmax <- function(params, rfac) {
                 is.numeric(rfac),
                 length(rfac) %in% c(1, nrow(params@species_params)),
                 all(rfac > 1))
-    if (params@rates_funcs$RDD != "srrNone") {
+    if (params@rates_funcs$RDD != "noRDD") {
         stop("setRmax can only be applied to params objects using the identity",
              " stock-recruitment function.")
     }
@@ -1171,7 +1171,7 @@ setRmax <- function(params, rfac) {
     params@species_params$R_max <- params@species_params$w_inf
     params@species_params$R_max <- (rfac - 1) * getRDI(params)
     
-    return(setReproduction(params, srr = "srrBevertonHolt"))
+    return(setReproduction(params, RDD = "BevertonHoltRDD"))
 }
 
 
@@ -1258,7 +1258,7 @@ steady <- function(params, t_max = 100, t_per = 7.5, tol = 10^(-2),
     
     # Force the recruitment to stay at the current level
     p@species_params$constant_recruitment <- getRDD(p)
-    p@rates_funcs$RDD <- "srrConstant"
+    p@rates_funcs$RDD <- "constantRDD"
     old_rdi <- getRDI(p)
     rdi_limit <- old_rdi / 1e7
     # Force other componens to stay at current level